refactor db-populate from 1 command into 3 separate commands, also `ruff format`
